+++ lib/amino_acid/amino_acid.dart
@@ -4,25 +4,16 @@ class AminoAcid {
   static const String className = 'AminoAcid';
   static const String version = '1';
   EnumAminoAcid type;
-  Map<String, String>? decoration;
-  Map<String, String>? replacement;
-  String? anotherName;
+  String? infoKey;
 
   /// * [type] : The amino acid type.
-  /// * [decoration] : Surface decoration. Key is target, Value is decorate object.
-  /// There is no fixed format. Format is free.
-  /// * [replacement] : Structure replacement information.
-  /// Key is target, Value is replacement object.
-  /// * [anotherName] : If there is a defined name for an amino acid with
-  /// some mutation added to a normal amino acid, it is convenient to describe it.
-  AminoAcid(this.type, {this.decoration, this.replacement, this.anotherName});
+  /// * [infoKey] : This amino acid information key.
+  /// The data are stored in amino acid sequence classes and referenced by this key.
+  AminoAcid(this.type, {this.infoKey});
 
   /// deep copy.
   AminoAcid deepCopy() {
-    return AminoAcid(type,
-        decoration: decoration != null ? {...decoration!} : null,
-        replacement: replacement != null ? {...replacement!} : null,
-        anotherName: anotherName);
+    return AminoAcid(type, infoKey: infoKey);
   }
 
   /// to map.
@@ -31,17 +22,13 @@ class AminoAcid {
     d['className'] = className;
     d['version'] = version;
     d['type'] = type.name;
-    d['decoration'] = decoration;
-    d['replacement'] = replacement;
-    d['anotherName'] = anotherName;
+    d['infoKey'] = infoKey;
     return d;
   }
 
   /// resume map.
   static AminoAcid fromDict(Map<String, dynamic> src) {
     return AminoAcid(EnumAminoAcid.values.byName(src['type']),
-        decoration: src['decoration'],
-        replacement: src['replacement'],
-        anotherName: src['anotherName']);
+        infoKey: src['infoKey']);
   }
 }