+++ lib/amino_acid/amino_acid_all.dart
@@ -1,4 +1,5 @@
 export 'amino_acid.dart';
+export 'amino_acid_info.dart';
 export 'amino_acid_sequence.dart';
 export 'enum_amino_acid.dart';
 export 'enum_amino_acid_sequence_direction.dart';